Use Chronicler for (most) games-by-day queries
blaseball.com's old data is missing pitcher names in a few places for
unknown reasons. Chronicler has it just fine in the cases we've seen, so
let's just use that.

There is an issue fetching the first 5 days of the Coffee Cup with
Chronicler currently, so continue using blaseball.com for that
tournament.